Consolidating CSS
